Remove deprecated code (#690)
* remove old qnns

* remove old kernel

* remove warning filtration

* fix tests

* fix style

* replace migration notebook with rst

* update reno

* skip statevector test on macos

* remove sampling neural network

* remove deprecations

* new tests

* new tests

* fix stestr

* replace bind_parameters with assign_parameters

* remove stestr fix
