+++ qiskit_machine_learning/kernels/trainable_fidelity_statevector_kernel.py
@@ -92,7 +92,7 @@ class TrainableFidelityStatevectorKernel(TrainableKernel, FidelityStatevectorKer
         self._feature_parameters = [
             parameter
             for parameter in feature_map.parameters
-            if parameter not in training_parameters
+            if parameter not in self._training_parameters
         ]
         self._parameter_dict = {parameter: None for parameter in self.feature_map.parameters}
