--- releasenotes/notes/deprecate-old-qk-qnn-af16b379ef3fc0c8.yaml
@@ -1,26 +1,0 @@
----
-deprecations:
-  - |
-    Since :mod:`qiskit.opflow` and :class:`~qiskit.utils.QuantumInstance` are deprecated in Qiskit,
-    Qiskit Machine Learning classes based on the deprecated Qiskit classes have been
-    deprecated as well:
-
-      - Class :class:`qiskit_machine_learning.neural_networks.CircuitQNN` is deprecated and
-        is superseded by :class:`qiskit_machine_learning.neural_networks.SamplerQNN`.
-
-      - Class :class:`qiskit_machine_learning.neural_networks.OpflowQNN` is deprecated and
-        is superseded by :class:`qiskit_machine_learning.neural_networks.EstimatorQNN`.
-
-      - Class :class:`qiskit_machine_learning.neural_networks.TwoLayerQNN` is deprecated and
-        has no direct replacement. Please make use of
-        :class:`qiskit_machine_learning.neural_networks.EstimatorQNN` instead.
-
-      - Class :class:`qiskit_machine_learning.kernels.QuantumKernel` is deprecated and
-        is superseded by :class:`qiskit_machine_learning.kernels.FidelityQuantumKernel`,
-        :class:`qiskit_machine_learning.kernels.TrainableFidelityQuantumKernel`, and
-        :class:`qiskit_machine_learning.kernels.FidelityStatevectorKernel`.
-
-    The above mentioned classes will be removed no sooner than 3 months after the release. Please
-    refer to the `migration guide <https://qiskit.org/ecosystem/machine-learning/migration/01_migration_guide_0.5.html>`__
-    on how to replace the deprecated classes with new ones.
-
