+++ test/algorithms/regressors/test_neural_network_regressor.py
@@ -1,6 +1,6 @@
 # This code is part of a Qiskit project.
 #
-# (C) Copyright IBM 2018, 2023.
+# (C) Copyright IBM 2022, 2023.
 #
 # This code is licensed under the Apache License, Version 2.0. You may
 # obtain a copy of this license in the LICENSE.txt file in the root directory
@@ -10,13 +10,13 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 """ Test Neural Network Regressor """
+from __future__ import annotations
+
 import itertools
 import os
 import tempfile
 import unittest
-import warnings
 from functools import partial
-from typing import Tuple
 
 from test import QiskitMachineLearningTestCase
 
@@ -25,14 +25,13 @@ from ddt import ddt, unpack, idata
 from qiskit.algorithms.optimizers import COBYLA, L_BFGS_B, SPSA
 from qiskit.circuit import Parameter, QuantumCircuit
 from qiskit.circuit.library import ZZFeatureMap, RealAmplitudes
-from qiskit.opflow import PauliSumOp
-from qiskit.utils import QuantumInstance, algorithm_globals, optionals
+from qiskit.utils import algorithm_globals
 from scipy.optimize import minimize
 
 from qiskit_machine_learning import QiskitMachineLearningError
 from qiskit_machine_learning.algorithms import SerializableModelMixin
 from qiskit_machine_learning.algorithms.regressors import NeuralNetworkRegressor
-from qiskit_machine_learning.neural_networks import TwoLayerQNN
+from qiskit_machine_learning.neural_networks import EstimatorQNN
 
 QUANTUM_INSTANCES = ["statevector", "qasm"]
 OPTIMIZERS = ["cobyla", "bfgs", "callable", None]
@@ -43,26 +42,11 @@ CALLBACKS = [True, False]
 class TestNeuralNetworkRegressor(QiskitMachineLearningTestCase):
     """Test Neural Network Regressor."""
 
-    @unittest.skipUnless(optionals.HAS_AER, "qiskit-aer is required to run this test")
     def setUp(self):
         super().setUp()
-        warnings.filterwarnings("ignore", category=DeprecationWarning)
 
         # specify quantum instances
         algorithm_globals.random_seed = 12345
-        from qiskit_aer import Aer
-
-        self.sv_quantum_instance = QuantumInstance(
-            Aer.get_backend("aer_simulator_statevector"),
-            seed_simulator=algorithm_globals.random_seed,
-            seed_transpiler=algorithm_globals.random_seed,
-        )
-        self.qasm_quantum_instance = QuantumInstance(
-            Aer.get_backend("aer_simulator"),
-            shots=100,
-            seed_simulator=algorithm_globals.random_seed,
-            seed_transpiler=algorithm_globals.random_seed,
-        )
 
         num_samples = 20
         eps = 0.2
@@ -73,13 +57,9 @@ class TestNeuralNetworkRegressor(QiskitMachineLearningTestCase):
         self.X = (ub - lb) * rng.random((num_samples, 1)) + lb
         self.y = np.sin(self.X[:, 0]) + eps * (2 * rng.random(num_samples) - 1)
 
-    def tearDown(self) -> None:
-        super().tearDown()
-        warnings.filterwarnings("always", category=DeprecationWarning)
-
     def _create_regressor(
-        self, opt, q_i, callback=None
-    ) -> Tuple[NeuralNetworkRegressor, TwoLayerQNN, QuantumCircuit]:
+        self, opt, callback=None
+    ) -> tuple[NeuralNetworkRegressor, EstimatorQNN, QuantumCircuit]:
         num_qubits = 1
 
         # construct simple feature map
@@ -92,12 +72,9 @@ class TestNeuralNetworkRegressor(QiskitMachineLearningTestCase):
         ansatz = QuantumCircuit(num_qubits, name="vf")
         ansatz.ry(param_y, 0)
 
-        if q_i == "statevector":
-            quantum_instance = self.sv_quantum_instance
-        elif q_i == "qasm":
-            quantum_instance = self.qasm_quantum_instance
-        else:
-            raise ValueError(f"Unsupported quantum instance: {q_i}")
+        qc = QuantumCircuit(num_qubits)
+        qc.compose(feature_map, inplace=True)
+        qc.compose(ansatz, inplace=True)
 
         if opt == "bfgs":
             optimizer = L_BFGS_B(maxiter=5)
@@ -109,27 +86,27 @@ class TestNeuralNetworkRegressor(QiskitMachineLearningTestCase):
             optimizer = None
 
         # construct QNN
-        regression_opflow_qnn = TwoLayerQNN(
-            num_qubits, feature_map, ansatz, quantum_instance=quantum_instance
+        regression_estimator_qnn = EstimatorQNN(
+            circuit=qc, input_params=feature_map.parameters, weight_params=ansatz.parameters
         )
 
         initial_point = np.zeros(ansatz.num_parameters)
 
         # construct the regressor from the neural network
         regressor = NeuralNetworkRegressor(
-            neural_network=regression_opflow_qnn,
+            neural_network=regression_estimator_qnn,
             loss="squared_error",
             optimizer=optimizer,
             initial_point=initial_point,
             callback=callback,
         )
 
-        return regressor, regression_opflow_qnn, ansatz
+        return regressor, regression_estimator_qnn, ansatz
 
-    @idata(itertools.product(OPTIMIZERS, QUANTUM_INSTANCES, CALLBACKS))
+    @idata(itertools.product(OPTIMIZERS, CALLBACKS))
     @unpack
-    def test_regressor_with_opflow_qnn(self, opt, q_i, cb_flag):
-        """Test Neural Network Regressor with Opflow QNN (Two Layer QNN)."""
+    def test_regressor_with_estimator_qnn(self, opt, cb_flag):
+        """Test Neural Network Regressor with Estimator QNN."""
         if cb_flag:
             history = {"weights": [], "values": []}
 
@@ -140,7 +117,7 @@ class TestNeuralNetworkRegressor(QiskitMachineLearningTestCase):
         else:
             callback = None
 
-        regressor, qnn, ansatz = self._create_regressor(opt, q_i, callback)
+        regressor, qnn, ansatz = self._create_regressor(opt, callback)
 
         # fit to data
         regressor.fit(self.X, self.y)
@@ -161,11 +138,10 @@ class TestNeuralNetworkRegressor(QiskitMachineLearningTestCase):
         np.testing.assert_array_equal(regressor.fit_result.x, regressor.weights)
         self.assertEqual(len(regressor.weights), ansatz.num_parameters)
 
-    @idata(itertools.product(OPTIMIZERS, QUANTUM_INSTANCES))
-    @unpack
-    def test_warm_start(self, opt, q_i):
+    @idata(OPTIMIZERS)
+    def test_warm_start(self, opt):
         """Test VQC when training from a warm start."""
-        regressor, _, _ = self._create_regressor(opt, q_i)
+        regressor, _, _ = self._create_regressor(opt)
         regressor.warm_start = True
 
         # Fit the regressor to the first half of the data.
@@ -188,12 +164,17 @@ class TestNeuralNetworkRegressor(QiskitMachineLearningTestCase):
         features = np.array([[0, 0], [0.1, 0.1], [0.4, 0.4], [1, 1]])
         labels = np.array([0, 0.1, 0.4, 1])
         num_inputs = 2
-        qnn = TwoLayerQNN(
-            num_inputs,
-            feature_map=ZZFeatureMap(num_inputs),
-            ansatz=RealAmplitudes(num_inputs),
-            observable=PauliSumOp.from_list([("Z" * num_inputs, 1)]),
-            quantum_instance=self.qasm_quantum_instance,
+
+        feature_map = ZZFeatureMap(num_inputs)
+        ansatz = RealAmplitudes(num_inputs)
+        qc = QuantumCircuit(num_inputs)
+        qc.compose(feature_map, inplace=True)
+        qc.compose(ansatz, inplace=True)
+
+        qnn = EstimatorQNN(
+            circuit=qc,
+            input_params=feature_map.parameters,
+            weight_params=ansatz.parameters,
         )
         regressor = NeuralNetworkRegressor(qnn, optimizer=COBYLA())
         regressor.fit(features, labels)
@@ -223,7 +204,8 @@ class TestNeuralNetworkRegressor(QiskitMachineLearningTestCase):
 
     def test_untrained(self):
         """Test untrained regressor."""
-        qnn = TwoLayerQNN(2)
+        qnn = EstimatorQNN(circuit=QuantumCircuit(2))
+
         regressor = NeuralNetworkRegressor(qnn)
         with self.assertRaises(QiskitMachineLearningError, msg="regressor.predict()"):
             regressor.predict(np.asarray([]))
@@ -236,7 +218,19 @@ class TestNeuralNetworkRegressor(QiskitMachineLearningTestCase):
 
     def test_callback_setter(self):
         """Test the callback setter."""
-        qnn = TwoLayerQNN(2, quantum_instance=self.qasm_quantum_instance)
+        num_inputs = 2
+        feature_map = ZZFeatureMap(num_inputs)
+        ansatz = RealAmplitudes(num_inputs)
+        qc = QuantumCircuit(num_inputs)
+        qc.compose(feature_map, inplace=True)
+        qc.compose(ansatz, inplace=True)
+
+        qnn = EstimatorQNN(
+            circuit=qc,
+            input_params=feature_map.parameters,
+            weight_params=ansatz.parameters,
+        )
+
         single_step_opt = SPSA(maxiter=1, learning_rate=0.01, perturbation=0.1)
         regressor = NeuralNetworkRegressor(qnn, optimizer=single_step_opt)
