+++ test/algorithms/regressors/test_qsvr.py
@@ -14,21 +14,16 @@
 import os
 import tempfile
 import unittest
-import warnings
 
 from test import QiskitMachineLearningTestCase
 
 import numpy as np
 
-from qiskit import BasicAer
 from qiskit.circuit.library import ZZFeatureMap
-from qiskit.utils import QuantumInstance, algorithm_globals
+from qiskit.utils import algorithm_globals
 from qiskit_machine_learning.algorithms import QSVR, SerializableModelMixin
-from qiskit_machine_learning.exceptions import (
-    QiskitMachineLearningError,
-    QiskitMachineLearningWarning,
-)
-from qiskit_machine_learning.kernels import QuantumKernel
+from qiskit_machine_learning.exceptions import QiskitMachineLearningWarning
+from qiskit_machine_learning.kernels import FidelityQuantumKernel
 
 
 class TestQSVR(QiskitMachineLearningTestCase):
@@ -36,16 +31,8 @@ class TestQSVR(QiskitMachineLearningTestCase):
 
     def setUp(self):
         super().setUp()
-        warnings.filterwarnings("ignore", category=DeprecationWarning)
 
         algorithm_globals.random_seed = 10598
-
-        self.statevector_simulator = QuantumInstance(
-            BasicAer.get_backend("statevector_simulator"),
-            shots=1,
-            seed_simulator=algorithm_globals.random_seed,
-            seed_transpiler=algorithm_globals.random_seed,
-        )
 
         self.feature_map = ZZFeatureMap(feature_dimension=2, reps=2)
 
@@ -62,49 +49,31 @@ class TestQSVR(QiskitMachineLearningTestCase):
         self.sample_test = np.asarray([[2.199114860, 5.15221195], [0.50265482, 0.06283185]])
         self.label_test = np.asarray([0, 1])
 
-    def tearDown(self) -> None:
-        super().tearDown()
-        warnings.filterwarnings("always", category=DeprecationWarning)
-
     def test_qsvr(self):
         """Test QSVR"""
-        qkernel = QuantumKernel(
-            feature_map=self.feature_map, quantum_instance=self.statevector_simulator
-        )
+        qkernel = FidelityQuantumKernel(feature_map=self.feature_map)
 
         qsvr = QSVR(quantum_kernel=qkernel)
         qsvr.fit(self.sample_train, self.label_train)
         score = qsvr.score(self.sample_test, self.label_test)
 
-        self.assertAlmostEqual(score, 0.38365, places=4)
-
-    def test_empty_kernel(self):
-        """Test QSVR with empty QuantumKernel"""
-        qkernel = QuantumKernel()
-        qsvr = QSVR(quantum_kernel=qkernel)
-
-        with self.assertRaises(QiskitMachineLearningError):
-            _ = qsvr.fit(self.sample_train, self.label_train)
+        self.assertAlmostEqual(score, 0.38359, places=4)
 
     def test_change_kernel(self):
         """Test QSVR with QuantumKernel later"""
-        qkernel = QuantumKernel(
-            feature_map=self.feature_map, quantum_instance=self.statevector_simulator
-        )
+        qkernel = FidelityQuantumKernel(feature_map=self.feature_map)
 
         qsvr = QSVR()
         qsvr.quantum_kernel = qkernel
         qsvr.fit(self.sample_train, self.label_train)
         score = qsvr.score(self.sample_test, self.label_test)
 
-        self.assertAlmostEqual(score, 0.38365, places=4)
+        self.assertAlmostEqual(score, 0.38359, places=4)
 
     def test_qsvr_parameters(self):
         """Test QSVR with extra constructor parameters"""
 
-        qkernel = QuantumKernel(
-            feature_map=self.feature_map, quantum_instance=self.statevector_simulator
-        )
+        qkernel = FidelityQuantumKernel(feature_map=self.feature_map)
 
         qsvr = QSVR(quantum_kernel=qkernel, tol=1e-4, C=0.5)
         qsvr.fit(self.sample_train, self.label_train)
@@ -127,9 +96,7 @@ class TestQSVR(QiskitMachineLearningTestCase):
         features = np.array([[0, 0], [0.1, 0.1], [0.4, 0.4], [1, 1]])
         labels = np.array([0, 0.1, 0.4, 1])
 
-        quantum_kernel = QuantumKernel(
-            feature_map=ZZFeatureMap(2), quantum_instance=self.statevector_simulator
-        )
+        quantum_kernel = FidelityQuantumKernel(feature_map=ZZFeatureMap(2))
         regressor = QSVR(quantum_kernel=quantum_kernel)
         regressor.fit(features, labels)
