+++ test/connectors/test_gpu_torch_connector.py
@@ -32,10 +32,6 @@ class TestGPUTorchConnector(QiskitMachineLearningTestCase, TestTorchConnector):
         else:
             self._device = torch.device("cuda")
 
-    def tearDown(self) -> None:
-        super().tearDown()
-        super().tear_down()
-
 
 if __name__ == "__main__":
     unittest.main()
