+++ test/connectors/test_gpu_torch_networks.py
@@ -31,10 +31,6 @@ class TestGPUTorchNetworks(QiskitMachineLearningTestCase, TestTorchNetworks):
         else:
             self._device = torch.device("cuda")
 
-    def tearDown(self) -> None:
-        super().tearDown()
-        super().tear_down()
-
 
 if __name__ == "__main__":
     unittest.main()
